From nam

+++ src/main/java/org/example/datn_website_supershoes/service/CartService.java
@@ -1,8 +1,13 @@
 package org.example.datn_website_supershoes.service;
 
 import org.example.datn_website_supershoes.Enum.Status;
+import org.example.datn_website_supershoes.dto.request.CartRequest;
+import org.example.datn_website_supershoes.dto.request.SizeRequest;
 import org.example.datn_website_supershoes.dto.response.CartResponse;
+import org.example.datn_website_supershoes.model.Account;
 import org.example.datn_website_supershoes.model.Cart;
+import org.example.datn_website_supershoes.model.Size;
+import org.example.datn_website_supershoes.repository.AccountRepository;
 import org.example.datn_website_supershoes.repository.CartRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,38 +21,41 @@ public class CartService {
 
     @Autowired
     private CartRepository cartRepository;
+    @Autowired
+    AccountRepository accountRepository;
 
-    public Cart createCart(Cart cart) {
-        return cartRepository.save(cart);
-    }
-
-    public List<CartResponse> getAllCarts() {
-        return cartRepository.listCartResponseByStatus(Status.ACTIVE.toString());
-    }
-
-    public Optional<Cart> getCartById(Long id) {
-        return cartRepository.findById(id);
-    }
-
-    public Cart updateCart(Long id, Cart cart) {
-        Cart existingCart = cartRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Cart not found"));
-
-        String[] ignoredProperties = {"id", "createdAt", "createdBy"};
-        BeanUtils.copyProperties(cart, existingCart, ignoredProperties);
-
-        if (cart.getAccount() != null) {
-            existingCart.setAccount(cart.getAccount());
+    public CartResponse getCartResponseByAccountId(Long accountId) {
+        Optional<CartResponse> cartResponse = cartRepository.CartResponse(accountId);
+        if (!cartResponse.isPresent()){
+            CartRequest cartRequest = new CartRequest();
+            cartRequest.setIdAccount(accountId);
+            Cart cart = cartRepository.save(convertCartRequestDTO(cartRequest));
+            CartResponse newCartResponse = new CartResponse();
+            newCartResponse.setId(cart.getId());
+            newCartResponse.setIdAccount(cart.getAccount().getId());
+            return newCartResponse;
+        }else{
+            return cartResponse.get();
         }
-        if (cart.getCartDetails() != null) {
-            existingCart.setCartDetails(cart.getCartDetails());
-        }
-
-        return cartRepository.save(cart);
     }
 
-    public void deleteCart(Long id) {
-        cartRepository.deleteById(id);
+    public void deleteCartById(Long id) {
+        if (cartRepository.existsById(id)) {
+            cartRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Cart với id " + id + " không tồn tại.");
+        }
+    }
+    public Cart convertCartRequestDTO(CartRequest cartRequest){
+        Account account = accountRepository.findById(cartRequest.getIdAccount())
+                .orElseThrow(() -> new RuntimeException("Không tìm tài khoản này!"));
+        Cart cart = Cart.builder()
+                .account(account)
+                .build();
+        cart.setCreatedBy(account.getName());
+        cart.setUpdatedBy(account.getName());
+        cart.setStatus(Status.ACTIVE.toString());
+        return cart;
     }
 }
